\name{print.ABF}
\alias{print.ABF}
\title{Prints a Short Description of an ABF Object}
\description{
Prints information about an \code{\link[=readABF]{ABF}} object and returns it invisibly. This helps to identify the dataset but does not overwhelm the user with the raw data.
}
\usage{
\S3method{print}{ABF}(x, ...)
}
\arguments{
  \item{x}{an object of class ABF, typically generated by \code{\link{readABF}}.}
  \item{...}{further arguments in case they are passed from other methods; they will be ignored.}
}
\examples{
\dontrun{
# loads "some_data.abf" and prints it
r <- readABF("some_data.abf")
print(r)
}
}
