\name{readABF-package}
\alias{readABF-package}
\docType{package}
\title{\packageTitle{readABF}}
\description{\packageDescription{readABF}}
\details{
This package aims to read both ABF2 files created by pClamp 10 and newer and ABF files created by pClamp 9 and older, especially ABF version 1.83.
\packageIndices{readABF}
}
\author{
\strong{Maintainer}: Stanislav Syekirin \email{syekirin@gmail.com}

Author: Stanislav Syekirin

Other contributors:
\itemize{
\item Florian Pein \email{fpein@uni-goettingen.de}
\item Deutsche Forschungsgemeinschaft SFB 803 Z2 [funder]
}

}
\keyword{package}