\name{abfReader-package}
\alias{abfReader-package}
\alias{abfReader}
\docType{package}
\title{\packageTitle{abfReader}}
\description{\packageDescription{abfReader}}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{abfReader}
\packageIndices{abfReader}
}
\author{
\packageAuthor{abfReader}

Maintainer: \packageMaintainer{abfReader}
}
\keyword{package}