\name{ABFReader-package}
\alias{ABFReader-package}
\alias{ABFReader}
\docType{package}
\title{
\packageTitle{ABFReader}
}
\description{
\packageDescription{ABFReader}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ABFReader}
\packageIndices{ABFReader}

}
\author{
\packageAuthor{ABFReader}

Maintainer: \packageMaintainer{ABFReader}
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
