\name{as.data.frame.ABF}
\alias{as.data.frame.ABF}
\title{Coerces an ABF Object to a Data Frame}

\description{
S3method to coerce an \code{\link[=readABF-package]{ABF}} object to a data frame. TODO tranformation to conductance
}

\usage{
\S3method{as.data.frame}{ABF}(x, row.names = NULL, optional = FALSE, ..., sweep = NULL, type = c("all", "conductance"),
current, voltage, unit = NULL)
}

\arguments{
  \item{x}{an object of class ABF, typically generated by \code{\link{readABF}}.}
  \item{row.names, optional, ...}{further arguments in case they are passed from other methods; will be ignored.}
  \item{sweep}{an integer, indicating which sweep should be coerced; required in case of multiple sweeps, otherwise ignored.}
  \item{type}{a string specifying the return value, see \sQuote{Values}.}
  \item{current}{an integer, indicating which channel contains the current; required if \code{type == "conductance"}, otherwise ignored.}
  \item{voltage}{an integer, indicating which channel contains the voltage; required if \code{type == "conductance"}, otherwise ignored.}
  \item{unit}{a string, specifying the unit of conductance, if \code{NULL} unit will be determined by current and voltage units in \code{x}; ignored if \code{type == "all"}}
}

\value{TODO: data.frame, time added}

\examples{
\dontrun{
# loads "some_data.abf" and coerces it to data frame
r <- readABF("some_data.abf")
as.data.frame(r)
}
}
