\name{plot.ABF}
\alias{plot.ABF}
\title{Plots an ABF Object}
\description{
S3 method to plot \code{\link[=readABF]{ABF}} objects. Converts the ABF object to a data frame, then plots the data frame.
}
\usage{
\S3method{plot}{ABF}(x, pch = ".", \dots, sweep = 1, type = c("all", "conductance"),
current, voltage, unit = NULL)
}
\arguments{
  \item{x}{an object of class ABF, typically generated by \code{\link{readABF}}.}
  \item{pch}{same as \code{pch} from \code{\link{par}}.}
  \item{...}{further arguments that will be passed to \code{\link{plot}}.}
  \item{sweep, type, current, voltage, unit}{arguments that will be passed to \code{\link{as.data.frame.ABF}}.}
}
\examples{
\dontrun{
# loads "some_data.abf" and plots it
r <- readABF("some_data.abf")
plot(r)
}
}
