\name{plot.ABF}
\alias{plot.ABF}
\title{Plots an ABF Objects}
\description{
S3method to plot \link[readABF]{ABF} objects.
}
\usage{
\S3method{plot}{ABF}(x, pch = ".", \dots, sweep = 1, type = c("all", "conductance"), current, voltage, unit = NULL)
}
\arguments{
  \item{x}{an object of class ABF, typically generated by \code{\link{readABF}}.}
  \item{pch}{todo}
  \item{...}{further arguments that will be passed to \code{\link{plot}}}
  \item{sweep, type, current, voltage, unit}{TODO link to as.data.frame.ABF}
}
\examples{
\dontrun{
# loads "some_data.abf" and plots it
r <- readABF("some_data.abf")
plot(r)
}
}
