\name{readABF-package}
\alias{readABF-package}
\docType{package}
\title{\packageTitle{readABF}}
\description{\packageDescription{readABF}}
\details{
Reads ABF2 files created by pClamp 10 and newer and ABF files created by pClamp 9 and older. Currently, code is tested under Linux only. For this purpose, multiple different data sets are used which we cannot include in the package. We would still appreciate to receive further data sets for testing purposes. A reporsitory to upload data sets can be provided at request.
\packageIndices{readABF}
}

\section{Acknowledgements}{Support of Deutsche Forschungsgemeinschaft grant SFB 803 Z2 is gratefully acknowledged.
\href{https://github.com/fcollman/abfload}{abfload.m} by Harald Hentschke, Forrest Collman and Ulrich Egert was an important source of knowledge about the inner workings of the format.
\href{https://cran.r-project.org/web/packages/abf2/index.html}{abf2} by Matthew Caldwell was a source of inspiration.
We would like to thank everyone who provided us with files so we could test our package with different format versions and operation modes.
Last but not least, the documentation provided by Molecular Devices on their site was helpful, too.
}

\keyword{package}
