\name{readABF-package}
\alias{readABF-package}
\docType{package}
\title{\packageTitle{readABF}}
\description{\packageDescription{readABF}}
\details{
Reads ABF2 files created by pClamp 10 and newer and ABF files created by pClamp 9 and older. Currently, code is tested under Linux only. For this purpose, multiple different data sets are used which we cannot include in the package. We would still appreciate to receive further data sets for testing purposes. A reporsitory to upload data sets can be provided at request.
\packageIndices{readABF}
}

\section{Acknowledgement}{Support of Deutsche Forschungsgemeinschaft grant SFB 803 Z2 is gratefully acknowledged.
TODO matlab, R package, abf testfiles}

\keyword{package}
