\name{readABF-package}
\alias{readABF-package}
\docType{package}
\title{\packageTitle{readABF}}
\description{\packageDescription{readABF}}
\details{
Reads ABF2 files, created by pClamp 10 and newer, and ABF files, created by pClamp 9 and older. Currently, code is tested under Linux only. For this purpose, multiple data sets are used which cannot be included in the package. We would very much appreciate to receive further data sets for testing purposes. A repository to upload data sets can be provided at request.
\packageIndices{readABF}
}

\section{Acknowledgements}{

\itemize{
\item Support of Deutsche Forschungsgemeinschaft grant SFB 803 Z2 is gratefully acknowledged.
\item \href{https://github.com/fcollman/abfload}{abfload} by Harald Hentschke, Forrest Collman and Ulrich Egert was an important source of information.
\item \href{https://CRAN.R-project.org/package=abf2}{abf2} by Matthew Caldwell was a useful starting point.
\item We would like to thank Harald Hentschke and his sources and the lab of Claudia Steinem for providing test data.
}
}

\keyword{package}
